+++ src/Yasumi/data/translations/pentecost.php
@@ -13,14 +13,15 @@
 // Translations for Whitsunday
 return [
     'da_DK' => 'Pinsedag',
-    'de_DE' => 'Pfingstsonntag',
     'de_AT' => 'Pfingsten',
     'de_CH' => 'Pfingsten',
+    'de_DE' => 'Pfingstsonntag',
     'el_GR' => 'Πεντηκοστή',
     'en_IE' => 'Whitsunday',
     'en_US' => 'Whitsunday',
     'et_EE' => 'Nelipühade 1. püha',
     'fi_FI' => 'Helluntaipäivä',
+    'fr_BE' => 'Pentecôte',
     'fr_CH' => 'Pentecôte',
     'ga_IE' => 'Domhnach Cincíse',
     'hu_HU' => 'Pünkösd',