+++ src/Yasumi/data/translations/secondChristmasDay.php
@@ -25,6 +25,8 @@ return [
     'et_EE' => 'Teine Jõulupüha',
     'fi_FI' => '2. joulupäivä',
     'hu_HU' => 'Karácsony másnapja',
+    'lt_LT' => 'Kalėdos (antra diena)',
+    'lv_LV' => 'Otrie Ziemassvētki',
     'nb_NO' => 'Andre juledag',
     'nl_BE' => 'Tweede Kerstdag',
     'nl_NL' => 'Tweede Kerstdag',