+++ src/Yasumi/data/translations/stJohnsDay.php
@@ -17,5 +17,7 @@ return [
     'es_ES' => 'Sant Joan',
     'et_EE' => 'Jaanipäev',
     'fi_FI' => 'Juhannuspäivä',
+    'lt_LT' => 'Rasos ir Joninių diena',
+    'lv_LV' => 'Jāņi',
     'sv_SE' => 'midsommardagen',
 ];