+++ tests/Japan/ChildrensDayTest.php
@@ -81,7 +81,7 @@ class ChildrensDayTest extends JapanBaseTestCase implements YasumiTestCaseInterf
             self::REGION,
             self::HOLIDAY,
             $this->generateRandomYear(self::ESTABLISHMENT_YEAR),
-            [self::LOCALE => '子供の日']
+            [self::LOCALE => 'こどもの日']
         );
     }