+++ tests/Lithuania/NewYearsDayTest.php
@@ -1,0 +1,70 @@
+<?php
+
+/**
+ * This file is part of the Yasumi package.
+ *
+ * Copyright (c) 2015 - 2017 AzuyaLabs
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Sacha Telgenhof <stelgenhof@gmail.com>
+ */
+
+namespace Yasumi\tests\Lithuania;
+
+use Yasumi\Holiday;
+use Yasumi\tests\YasumiTestCaseInterface;
+
+/**
+ * Class containing tests for New Year's Day in Lithuania.
+ *
+ * @author Gedas Lukošius <gedas@lukosius.me>
+ */
+class NewYearsDayTest extends LithuaniaBaseTestCase implements YasumiTestCaseInterface
+{
+    /**
+     * The name of the holiday to be tested
+     */
+    const HOLIDAY = 'newYearsDay';
+
+    /**
+     * @return array
+     */
+    public function holidayDataProvider()
+    {
+        return $this->generateRandomDates(1, 1, self::TIMEZONE);
+    }
+
+    /**
+     * @dataProvider holidayDataProvider
+     *
+     * @param int       $year
+     * @param \DateTime $expected
+     */
+    public function testHoliday($year, \DateTime $expected)
+    {
+        $this->assertHoliday(self::REGION, self::HOLIDAY, $year, $expected);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function testTranslation()
+    {
+        $this->assertTranslatedHolidayName(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(),
+            [self::LOCALE => 'Naujųjų metų diena']
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function testHolidayType()
+    {
+        $this->assertHolidayType(self::REGION, self::HOLIDAY, $this->generateRandomYear(), Holiday::TYPE_OFFICIAL);
+    }
+}